# Fix gridding, allow gridding based on functions.