--- a/src/pint/gridutils.py
+++ b/src/pint/gridutils.py
@@ -3,6 +3,7 @@ import copy
 import multiprocessing
 import os
 from multiprocessing import Process, Queue
+from pathos.multiprocessing import ProcessingPool as Pool
 
 import astropy.constants as const
 import astropy.units as u
@@ -13,8 +14,9 @@ import pint.utils
 __all__ = ["grid_chisq", "grid_chisq_mp", "plot_grid_chisq"]
 
 
-def _grid_docol(ftr, par1_name, par1, par2_name, par2_grid, ii, q):
+def _grid_docol(ftr, par1_name, par1, par2_name, par2_grid):
     """Worker process that computes one row of the chisq grid"""
+    chisq = np.zeros(len(par2_grid))
     for jj, par2 in enumerate(par2_grid):
         # Make a full copy of the fitter to work with
         myftr = copy.deepcopy(ftr)
@@ -24,29 +26,37 @@ def _grid_docol(ftr, par1_name, par1, par2_name, par2_grid, ii, q):
         getattr(myftr.model, par2_name).frozen = True
         getattr(myftr.model, par1_name).quantity = par1
         getattr(myftr.model, par2_name).quantity = par2
-        chisq = myftr.fit_toas()
-        print(".", end="")
-        q.put([ii, jj, chisq])
+        chisq[jj] = myftr.fit_toas()
+    return chisq
 
 
-def grid_chisq_mp(ftr, par1_name, par1_grid, par2_name, par2_grid, ncpu=None):
+def _grid_doonefit(ftr, parnames, parvalues):
+    """Worker process that computes one fit for the chisq grid"""
+    # Make a full copy of the fitter to work with
+    myftr = copy.deepcopy(ftr)
+    for parname, parvalue in zip(parnames, parvalues):
+        # Freeze the  params we are going to grid over and set their values
+        # All other unfrozen parameters will be fitted for at each grid point
+        getattr(myftr.model, parname).frozen = True
+        getattr(myftr.model, parname).quantity = parvalue
+    return myftr.fit_toas()
+
+
+def grid_chisq_mp(ftr, parnames, parvalues, ncpu=None):
     """Compute chisq over a grid of two parameters, multiprocessing version
 
-    Use Python's multiprocessing package to do a parallel computation of
-    chisq over 2-D grid of parameters.
+    Use pathos's multiprocessing package to do a parallel computation of
+    chisq over 2-D grid of parameters.  Need this instead of stock python because
+    of unpicklable objects.
 
     Parameters
     ----------
     ftr
         The base fitter to use.
-    par1_name : str
-        Name of the first parameter to grid over
-    par1_grid : array, Quantity
-        Array of par1 values for column of the output matrix
-    par2_name : str
-        Name of the second parameter to grid over
-    par2_grid : array, Quantity
-        Array of par2 values for column of the output matrix
+    parnames : list
+        Names of the parameters to grid over
+    parvalues : list
+        List of parameter values to grid over (each should be array of Quantity)
     ncpu : int, optional
         Number of processes to use in parallel. Default is number of CPUs available
 
@@ -59,47 +69,25 @@ def grid_chisq_mp(ftr, par1_name, par1_grid, par2_name, par2_grid, ncpu=None):
         # Use al available CPUs
         ncpu = multiprocessing.cpu_count()
 
-    # Instantiate a Queue for getting return values from the worker processes
-    q = Queue()
+    pool = Pool(ncpu)
+    out = np.meshgrid(*parvalues)
+    chi2 = np.zeros(out[0].shape)
+    it = np.nditer(out[0], flags=["multi_index"])
 
-    chi2 = np.zeros((len(par1_grid), len(par2_grid)))
-    # First create all the processes and put them in a list
-    processes = []
-    # Want par1 on X-axis and par2 on y-axis
-    for ii, par1 in enumerate(par1_grid):
-        # ii indexes rows, now for have each column done by a different process
-        proc = Process(
-            target=_grid_docol, args=(ftr, par1_name, par1, par2_name, par2_grid, ii, q)
-        )
-        processes.append(proc)
-
-    # Now consume the list of processes by starting up to ncpu processes at a time
-    while len(processes):
-        # Start up to ncpu processes
-        started = []
-        for cpunum in range(ncpu):
-            if len(processes):
-                proc = processes.pop()
-                proc.start()
-                started.append(proc)
-        # Collect all the results from the started processes
-        for proc in started * len(par2_grid):
-            ii, jj, ch = q.get()
-            # Array index here is rownum, colnum so translates to y, x
-            chi2[jj, ii] = ch
-
-        # Now join each of those that are done to close them out
-        # This will be inefficient if the processes have large differences in runtime, since there
-        # is a synchronization point. In this case it should not be a problem.
-        for proc in started:
-            proc.join()
-            print("|", end="")
-        print("")
+    # First create all the processes and put them in a pool
+    results = pool.map(
+        _grid_doonefit,
+        (ftr,) * len(out[0].flatten()),
+        (parnames,) * len(out[0].flatten()),
+        list(zip(*[x.flatten() for x in out])),
+    )
+    for j, x in enumerate(it):
+        chi2[it.multi_index] = results[j]
 
     return chi2
 
 
-def grid_chisq(ftr, par1_name, par1_grid, par2_name, par2_grid, printprogress=True):
+def grid_chisq(ftr, parnames, parvalues, printprogress=True):
     """Compute chisq over a grid of two parameters, serial version
 
     Single-threaded computation of chisq over 2-D grid of parameters.
@@ -108,20 +96,16 @@ def grid_chisq(ftr, par1_name, par1_grid, par2_name, par2_grid, printprogress=Tr
     ----------
     ftr
         The base fitter to use.
-    par1_name : str
-        Name of the first parameter to grid over
-    par1_grid : array, Quantity
-        Array of par1 values for column of the output matrix
-    par2_name : str
-        Name of the second parameter to grid over
-    par2_grid : array, Quantity
-        Array of par2 values for column of the output matrix
+    parnames : list
+        Names of the parameters to grid over
+    parvalues : list
+        List of parameter values to grid over (each should be array of Quantity)
     printprogress : bool, optional
         Print indications of progress
 
     Returns
     -------
-    array : 2-D array of chisq values with par1 varying in columns and par2 varying in rows
+    array : 2-D array of chisq values 
 
     """
 
@@ -130,24 +114,23 @@ def grid_chisq(ftr, par1_name, par1_grid, par2_name, par2_grid, printprogress=Tr
     gridmod = copy.deepcopy(ftr.model)
     ftr.model = gridmod
 
-    # Freeze the two params we are going to grid over
-    getattr(ftr.model, par1_name).frozen = True
-    getattr(ftr.model, par2_name).frozen = True
+    # Freeze the  params we are going to grid over
+    for parname in parnames:
+        getattr(ftr.model, parname).frozen = True
 
     # All other unfrozen parameters will be fitted for at each grid point
-
-    chi2 = np.zeros((len(par2_grid), len(par1_grid)))
-    # Want par1 on X-axis and par2 on y-axis
-    for ii, par1 in enumerate(par1_grid):
-        getattr(ftr.model, par1_name).quantity = par1
-        for jj, par2 in enumerate(par2_grid):
-            getattr(ftr.model, par2_name).quantity = par2
-            # Array index here is rownum, colnum so translates to y, x
-            chi2[jj, ii] = ftr.fit_toas()
-            if printprogress:
-                print(".", end="")
+    out = np.meshgrid(*parvalues)
+    chi2 = np.zeros(out[0].shape)
+    it = np.nditer(out[0], flags=["multi_index"])
+    for x in it:
+        for parnum, parname in enumerate(parnames):
+            getattr(ftr.model, parname).quantity = out[parnum][it.multi_index]
+        chi2[it.multi_index] = ftr.fit_toas()
         if printprogress:
-            print("")
+            print(".", end="")
+
+    if printprogress:
+        print("")
 
     # Restore saved model
     ftr.model = savemod