Still very much WIP (no tests, need to work on MP versions).  Fixes #1135 but also does more.

For instance, previously `gridutils` could only define a grid corresponding to a 1d range of 2 parameters.  But what if you wanted a uniform grid over another parameter (cos(i) instead of sin(i))?  Or something involving more than one quantity?

I'm trying to add the capability to define the grid and the parameters separately, where the parameters are functions of the grid.  So:

```
pint.gridutils.grid_chisq_derived(fsave,
                                                'M2',
                                                lambda x,y: x,
                                                M2,
                                                'SINI',
                                                lambda x,y: np.sqrt(1-y**2),
                                                cosi)
```
defines a grid in cos(i) but then uses that to compute SINI.

This can be extended to more complicated parameterizations (e.g., define a grid in M1, but then compute M2 using some PK parameters).  There isn't a huge advantage to doing this for the simple version of the function (you can easily roll your own), but with the MP versions it may be worth it.