@paulray 2021-08-30
Great, sounds like a good extension of the gridding utilities!