@dlakaplan 2021-08-30
Actually, I can't even run the MP version on the data I'm using for testing because the `maskParameter` needed for a jump isn't picklable (it includes lambda's).  Has anybody used that function (`grid_chisq_mp`)?  I know there are ways around this (use `dill` instead of `pickle`) but I don't want to reinvent something that had already been solved.