@dlakaplan 2021-08-30
I'm not sure this is the best way, but I have a version that works for MP now (using pathos).  

I also changed the grid behavior so that it can accept an arbitrary number of parameters to grid over