@dlakaplan 2021-08-30
Now the interface looks like:
```
pint.gridutils.grid_chisq(fsave, ('M2','SINI'),
                                       (M2,sini))
```
so you can have 1 or 2 or 3 or more parameters to grid over.