@aarchiba 2021-09-01
The answer is we need to fix #1076 so that PINT models can be pickled. It's not a particularly hard fix, I was just deferring it until we could fix it as part of #1071, but it seems to be more urgent than that.