@aarchiba 2021-09-01
For multiprocessing, one tool I have found helpful is to use python's `concurrent.futures` model and have parallel functions accept an Executor object - this way someone can pass in a multiprocessing executor with an appropriate number of workers and your code will use it, or a multithreading executor if that's useful, or some distributed-computing executor; Executors provide a standardized interface to parallel processing that's powerful enough for what the gridutils want.

https://docs.python.org/3/library/concurrent.futures.html