@dlakaplan 2021-09-01
Some updated timing:
```
old algorithm grid [5, 5] 1 cpus in 1.5876991748809814 s
old algorithm grid [5, 5] 4 cpus in 0.9443788528442383 s
new algorithm grid [5, 5] 1 cpus in 1.6766939163208008 s
new algorithm grid [5, 5] 4 cpus in 1.529547929763794 s
old algorithm grid [10, 10] 1 cpus in 6.481395959854126 s
old algorithm grid [10, 10] 4 cpus in 4.896639108657837 s
new algorithm grid [10, 10] 1 cpus in 7.438166856765747 s
new algorithm grid [10, 10] 4 cpus in 3.938685178756714 s
old algorithm grid [20, 20] 1 cpus in 31.33030915260315 s
old algorithm grid [20, 20] 4 cpus in 18.0501708984375 s
new algorithm grid [20, 20] 1 cpus in 27.26108193397522 s
new algorithm grid [20, 20] 4 cpus in 22.748576879501343 s
old algorithm grid [50, 50] 1 cpus in 172.6981749534607 s
old algorithm grid [50, 50] 4 cpus in 114.43707799911499 s
new algorithm grid [50, 50] 1 cpus in 170.2247428894043 s
new algorithm grid [50, 50] 4 cpus in 118.5853750705719 s
old algorithm grid [150, 150] 1 cpus in 1669.5488622188568 s
old algorithm grid [150, 150] 4 cpus in 948.948086977005 s
new algorithm grid [150, 150] 1 cpus in 1533.0161559581757 s
new algorithm grid [150, 150] 4 cpus in 943.6808099746704 s
new algorithm grid [5, 5, 5] 1 cpus in 7.619338035583496 s
new algorithm grid [5, 5, 5] 4 cpus in 5.0489091873168945 s
new algorithm grid [10, 10, 10] 1 cpus in 64.79371809959412 s
new algorithm grid [10, 10, 10] 4 cpus in 33.99533987045288 s
new algorithm grid [20, 20, 20] 1 cpus in 693.8935396671295 s
new algorithm grid [20, 20, 20] 4 cpus in 418.54967498779297 s
new algorithm grid [50, 50, 50] 1 cpus in 8233.895636081696 s
new algorithm grid [50, 50, 50] 4 cpus in 4726.437324762344 s
```
Harder to parse, but the new (cell-by-cell) vs. old (column) parallelization seems pretty comparable for the larger grids.

I have yet to change to the more generic Executor objects, though.