@dlakaplan 2021-09-01
I re-wrote the MP stuff to use Executors, as Anne suggested.  And I cleaned up the interface a bit. There is still some old stuff and some WIP stuff in the end of the module (for continued testing) but I think this is getting better.

Note that I wrote an example, but I don't think it's getting build since the example notebooks are still not being listed properly (#1137?)