@aarchiba 2021-09-02
Notebooks are not automatically listed, if you want it in the documentation you have to list it with the others (under tutorials.rst I think). (Which is not to say that our notebooks might not also be broken.)