@aarchiba 2021-09-02
> This build is failing because some of the tests are failing. This is related to the change to use Executors: it works sometimes, but when I try it with prefix parameters I _think_ it's hanging because of an unpicklability issue. I don't know why it's hanging (which leads to tests being canceled) rather than crashing. The functions work with other parameters but clearly we need to solve this. So should probably delay this until after #1076 is solved.

I added comments on that issue on the easiest way to fix that. I won't be able to get to that for a week or two at the least.