@dlakaplan 2021-09-02
Thanks.  It may be that we can push forward with this PR and just include a cautionary note that MP won't work with all models.  I also don't have a lot of time to start something new.