@aarchiba 2021-09-02
Easiest in that case is to xfail the problem tests with an explanation involving pickling.