@dlakaplan 2021-09-02
I have tried that, with a pytest timeout and a pytest fail.  That in itself works OK.  But the test still doesn't return - it doesn't give a pickle error like the other modules did, but just sits there.  So maybe better to do without that test.