@dlakaplan 2021-09-08
Incorporates #1140 to enable use of pickled parameters.

Usage is now:
`grid_chisq(..., executor=, ncpu=,...)`
With the following behavior:
| `executor` | `ncpu` | result |
| ----------- | ------- | ------ |
| existing object | any | uses existing `executor` |
| None | 1 | uses single-processor |
| None | None | creates default `executor` with `cpu_count` workers |
| None | >1 | creates default `executor` with desired number of workers |

Does that make sense?  

Finally, I changed the test to use  2 CPUs (for the multi-CPU tests)