@dlakaplan 2021-09-08
I put a table like the above in the doctoring.  On my local build it looks fine, but once again it isn't rendering properly at https://nanograv-pint--1136.org.readthedocs.build/en/1136/_autosummary/pint.gridutils.grid_chisq.html

(#1099)