@dlakaplan 2021-09-13
@aarchiba (or others) : do you know of appropriate Executor implementations for HPC environments?  Our local cluster is scheduled with slurm, but I can't find an obvious interface to use that.