@scottransom 2021-09-13
@dlakaplan For most clusters I've seen, the scheduler is separate from the program execution.  So I would expect that you'd use a normal Python+MPI executor, which you would then call from your job script which slurm would schedule.