@dlakaplan 2021-09-13
Current progress bar uses `tqdm` which is optional (#1144)